Fix folder structure back to restructure format
