+++ src/components/modules/banners/ausMomentThankYouBanner/AusMomentThankYouBanner.tsx
@@ -12,8 +12,8 @@ import {
 } from '@guardian/src-icons';
 import { from } from '@guardian/src-foundations/mq';
 import { BannerProps } from '../BannerTypes';
-import { addTrackingParams } from '../../../../../lib/tracking';
-import { setContributionsBannerClosedTimestamp } from '../localStorage';
+import { addTrackingParams } from '../../../../lib/tracking';
+import { setContributionsBannerClosedTimestamp } from '../ausMomentContributionsBanner/localStorage';
 
 const banner = css`
     display: flex;
